# data-mannagement/data-mannagement/api/scenemanage/v1/scene_manage_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.6.2
// - protoc             v5.29.3
// source: scenemanage/v1/scene_manage.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationSceneManageCopyScene = "/api.scenemanage.v1.SceneManage/CopyScene"
const OperationSceneManageCreateScene = "/api.scenemanage.v1.SceneManage/CreateScene"
const OperationSceneManageDeleteSceneItem = "/api.scenemanage.v1.SceneManage/DeleteSceneItem"
const OperationSceneManageEditScene = "/api.scenemanage.v1.SceneManage/EditScene"
const OperationSceneManageEditSceneDisabled = "/api.scenemanage.v1.SceneManage/EditSceneDisabled"
const OperationSceneManageSceneDetail = "/api.scenemanage.v1.SceneManage/SceneDetail"
const OperationSceneManageSceneTemplateList = "/api.scenemanage.v1.SceneManage/SceneTemplateList"

type SceneManageHTTPServer interface {
	// CopyScene复制场景
	CopyScene(context.Context, *CopySceneRequest) (*CopySceneReply, error)
	// CreateScene创建场景
	CreateScene(context.Context, *CreateSceneRequest) (*CreateSceneReply, error)
	// DeleteSceneItem 删除场景条目
	DeleteSceneItem(context.Context, *DeleteSceneItemRequest) (*emptypb.Empty, error)
	// EditScene 编辑场景(包括编辑名字，新增场景item，默认为可用)
	EditScene(context.Context, *EditSceneRequest) (*emptypb.Empty, error)
	// EditSceneDisabled 编辑场景模版状态
	EditSceneDisabled(context.Context, *EditSceneDisabledRequest) (*emptypb.Empty, error)
	// SceneDetail 场景详情
	SceneDetail(context.Context, *SceneDetailRequest) (*SceneDetailReplay, error)
	// SceneTemplateList 场景模版列表
	SceneTemplateList(context.Context, *SceneTemplateListRequest) (*SceneTemplateListReply, error)
}

func RegisterSceneManageHTTPServer(s *http.Server, srv SceneManageHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/data/scene/create", _SceneManage_CreateScene0_HTTP_Handler(srv))
	r.POST("/v1/data/scene/copy", _SceneManage_CopyScene0_HTTP_Handler(srv))
	r.POST("/v1/data/scene/edit/status", _SceneManage_EditSceneDisabled0_HTTP_Handler(srv))
	r.POST("/v1/data/scene/edit", _SceneManage_EditScene0_HTTP_Handler(srv))
	r.GET("/v1/data/scene/detail", _SceneManage_SceneDetail0_HTTP_Handler(srv))
	r.GET("/v1/data/scene/list", _SceneManage_SceneTemplateList0_HTTP_Handler(srv))
	r.DELETE("/v1/data/scene/item/delete", _SceneManage_DeleteSceneItem0_HTTP_Handler(srv))
}

func _SceneManage_CreateScene0_HTTP_Handler(srv SceneManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateSceneRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSceneManageCreateScene)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateScene(ctx, req.(*CreateSceneRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateSceneReply)
		return ctx.Result(200, reply)
	}
}

func _SceneManage_CopyScene0_HTTP_Handler(srv SceneManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CopySceneRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSceneManageCopyScene)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CopyScene(ctx, req.(*CopySceneRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CopySceneReply)
		return ctx.Result(200, reply)
	}
}

func _SceneManage_EditSceneDisabled0_HTTP_Handler(srv SceneManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in EditSceneDisabledRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSceneManageEditSceneDisabled)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.EditSceneDisabled(ctx, req.(*EditSceneDisabledRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _SceneManage_EditScene0_HTTP_Handler(srv SceneManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in EditSceneRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSceneManageEditScene)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.EditScene(ctx, req.(*EditSceneRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _SceneManage_SceneDetail0_HTTP_Handler(srv SceneManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SceneDetailRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSceneManageSceneDetail)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SceneDetail(ctx, req.(*SceneDetailRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SceneDetailReplay)
		return ctx.Result(200, reply)
	}
}

func _SceneManage_SceneTemplateList0_HTTP_Handler(srv SceneManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SceneTemplateListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSceneManageSceneTemplateList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SceneTemplateList(ctx, req.(*SceneTemplateListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SceneTemplateListReply)
		return ctx.Result(200, reply)
	}
}

func _SceneManage_DeleteSceneItem0_HTTP_Handler(srv SceneManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteSceneItemRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSceneManageDeleteSceneItem)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteSceneItem(ctx, req.(*DeleteSceneItemRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

type SceneManageHTTPClient interface {
	CopyScene(ctx context.Context, req *CopySceneRequest, opts ...http.CallOption) (rsp *CopySceneReply, err error)
	CreateScene(ctx context.Context, req *CreateSceneRequest, opts ...http.CallOption) (rsp *CreateSceneReply, err error)
	DeleteSceneItem(ctx context.Context, req *DeleteSceneItemRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	EditScene(ctx context.Context, req *EditSceneRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	EditSceneDisabled(ctx context.Context, req *EditSceneDisabledRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	SceneDetail(ctx context.Context, req *SceneDetailRequest, opts ...http.CallOption) (rsp *SceneDetailReplay, err error)
	SceneTemplateList(ctx context.Context, req *SceneTemplateListRequest, opts ...http.CallOption) (rsp *SceneTemplateListReply, err error)
}

type SceneManageHTTPClientImpl struct {
	cc *http.Client
}

func NewSceneManageHTTPClient(client *http.Client) SceneManageHTTPClient {
	return &SceneManageHTTPClientImpl{client}
}

func (c *SceneManageHTTPClientImpl) CopyScene(ctx context.Context, in *CopySceneRequest, opts ...http.CallOption) (*CopySceneReply, error) {
	var out CopySceneReply
	pattern := "/v1/data/scene/copy"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSceneManageCopyScene))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *SceneManageHTTPClientImpl) CreateScene(ctx context.Context, in *CreateSceneRequest, opts ...http.CallOption) (*CreateSceneReply, error) {
	var out CreateSceneReply
	pattern := "/v1/data/scene/create"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSceneManageCreateScene))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *SceneManageHTTPClientImpl) DeleteSceneItem(ctx context.Context, in *DeleteSceneItemRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/v1/data/scene/item/delete"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSceneManageDeleteSceneItem))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *SceneManageHTTPClientImpl) EditScene(ctx context.Context, in *EditSceneRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/v1/data/scene/edit"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSceneManageEditScene))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *SceneManageHTTPClientImpl) EditSceneDisabled(ctx context.Context, in *EditSceneDisabledRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/v1/data/scene/edit/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSceneManageEditSceneDisabled))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *SceneManageHTTPClientImpl) SceneDetail(ctx context.Context, in *SceneDetailRequest, opts ...http.CallOption) (*SceneDetailReplay, error) {
	var out SceneDetailReplay
	pattern := "/v1/data/scene/detail"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSceneManageSceneDetail))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *SceneManageHTTPClientImpl) SceneTemplateList(ctx context.Context, in *SceneTemplateListRequest, opts ...http.CallOption) (*SceneTemplateListReply, error) {
	var out SceneTemplateListReply
	pattern := "/v1/data/scene/list"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSceneManageSceneTemplateList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}