# data-mannagement/data-mannagement/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/data/scene/copy:
        post:
            tags:
                - SceneManage
            description: 复制场景
            operationId: SceneManage_CopyScene
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.scenemanage.v1.CopySceneRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.scenemanage.v1.CopySceneReply'
    /v1/data/scene/create:
        post:
            tags:
                - SceneManage
            description: 创建场景
            operationId: SceneManage_CreateScene
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.scenemanage.v1.CreateSceneRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.scenemanage.v1.CreateSceneReply'
    /v1/data/scene/detail:
        get:
            tags:
                - SceneManage
            description: 场景详情
            operationId: SceneManage_SceneDetail
            parameters:
                - name: sceneTemplateId
                  in: query
                  description: 模版场景id
                  schema:
                    type: integer
                    format: uint64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.scenemanage.v1.SceneDetailReplay'
    /v1/data/scene/edit:
        post:
            tags:
                - SceneManage
            description: 编辑场景(包括编辑名字，新增场景item，默认为可用)
            operationId: SceneManage_EditScene
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.scenemanage.v1.EditSceneRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /v1/data/scene/edit/status:
        post:
            tags:
                - SceneManage
            description: 编辑场景模版状态
            operationId: SceneManage_EditSceneDisabled
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.scenemanage.v1.EditSceneDisabledRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /v1/data/scene/item/delete:
        delete:
            tags:
                - SceneManage
            description: 删除场景条目
            operationId: SceneManage_DeleteSceneItem
            parameters:
                - name: sceneItemId
                  in: query
                  description: 条目id
                  schema:
                    type: integer
                    format: uint64
            responses:
                "200":
                    description: OK
                    content: {}
    /v1/data/scene/list:
        get:
            tags:
                - SceneManage
            description: 场景模版列表
            operationId: SceneManage_SceneTemplateList
            parameters:
                - name: pageSize
                  in: query
                  description: 每页条数（必填）
                  schema:
                    type: integer
                    format: uint64
                - name: pageNumber
                  in: query
                  description: 页码数,从1开始（必填）
                  schema:
                    type: integer
                    format: uint64
                - name: sortBy
                  in: query
                  description: 排序字段（非必须）
                  schema:
                    type: string
                - name: order
                  in: query
                  description: 排序方式("asc" or "desc") （非必须）
                  schema:
                    type: string
                - name: searchSceneName
                  in: query
                  description: 按照场景名称进行模糊搜索（非必须）
                  schema:
                    type: string
                - name: searchSceneDisabled
                  in: query
                  description: 按照是否禁用搜索, all：搜索全部，disabled：搜索禁用的 enabled：搜索未禁用的
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.scenemanage.v1.SceneTemplateListReply'
...
components:
    schemas:
...
        api.scenemanage.v1.AddSceneItem:
            type: object
            properties:
                id:
                    type: integer
                    description: 场景id,新增就是0，否则就是原来id
                    format: uint64
                sceneItemName:
                    type: string
                    description: 场景条目名字
                type:
                    type: string
                    description: '类型，dict:字典类型，text: 文本，integer:整数，decimal:小数，datetime:日期，picture:图片，video:视频，audio:音频，file:文件'
                dictTemplateId:
                    type: integer
                    description: 字典模版id,如果dict类型为dict,必填
                    format: uint64
                multiValue:
                    type: boolean
                    description: 是否可写多个值，0:否 1:是
                isHide:
                    type: boolean
                    description: 是否隐藏，0:否 1:是
                isRequired:
                    type: boolean
                    description: 是否必填，0:否 1:是
                isSupportSearch:
                    type: boolean
                    description: 是否必填，0:否 1:是
                sortNumber:
                    type: integer
                    description: item排列顺序，从1开始
                    format: uint64
            description: 增加的场景条目
        api.scenemanage.v1.CopySceneReply:
            type: object
            properties:
                sceneTemplateId:
                    type: integer
                    description: 模版场景id
                    format: uint64
        api.scenemanage.v1.CopySceneRequest:
            type: object
            properties:
                sceneTemplateId:
                    type: integer
                    description: 模版场景id
                    format: uint64
                sceneName:
                    type: string
                    description: 新场景名字
        api.scenemanage.v1.CreateSceneReply:
            type: object
            properties:
                sceneTemplateId:
                    type: integer
                    description: 场景模版id
                    format: uint64
        api.scenemanage.v1.CreateSceneRequest:
            type: object
            properties:
                sceneName:
                    type: string
                    description: 场景名字
                sceneItem:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.scenemanage.v1.AddSceneItem'
        api.scenemanage.v1.EditSceneDisabledRequest:
            type: object
            properties:
                sceneTemplateId:
                    type: integer
                    description: 场景模版id
                    format: uint64
                isDisabled:
                    type: boolean
                    description: 是否禁用, 否：0，禁用：1，默认0
        api.scenemanage.v1.EditSceneRequest:
            type: object
            properties:
                sceneTemplateId:
                    type: integer
                    description: 模版场景id
                    format: uint64
                sceneName:
                    type: string
                    description: 场景名字
                sceneItem:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.scenemanage.v1.AddSceneItem'
                    description: 增加的场景条目
        api.scenemanage.v1.SceneDetailReplay:
            type: object
            properties:
                sceneItem:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.scenemanage.v1.SceneDetailReplay_SceneItem'
                    description: 增加的场景条目
                sceneTemplateDetail:
                    $ref: '#/components/schemas/api.scenemanage.v1.SceneTemplate'
        api.scenemanage.v1.SceneDetailReplay_SceneItem:
            type: object
            properties:
                id:
                    type: integer
                    description: 条目id
                    format: uint64
                type:
                    type: string
                    description: '类型，dict:字典类型，text: 文本，integer:整数，decimal:小数，datetime:日期，picture:图片，video:视频，audio:音频，file:文件'
                dictTemplateId:
                    type: integer
                    description: 字典模版id,如果dict类型为dict
                    format: uint64
                dictTemplateName:
                    type: string
                    description: 字典模版名字,如果dict类型为dict
                multiValue:
                    type: boolean
                    description: 是否可写多个值，0:否 1:是
                isHide:
                    type: boolean
                    description: 是否隐藏，0:否 1:是
                isRequired:
                    type: boolean
                    description: 是否必填，0:否 1:是
                isSupportSearch:
                    type: boolean
                    description: 是否支持搜索，0:否 1:是
                sortNumber:
                    type: integer
                    description: item排列顺序，从1开始
                    format: uint64
                sceneItemName:
                    type: string
                    description: 场景条目名字
        api.scenemanage.v1.SceneTemplate:
            type: object
            properties:
                sceneTemplateId:
                    type: integer
                    description: 模版场景id
                    format: uint64
                sceneName:
                    type: string
                    description: 场景名字
                sceneIsDisabled:
                    type: boolean
                    description: 是否禁用, 否：0，禁用：1，默认0
                sceneIsUsed:
                    type: boolean
                    description: 是否使用中，0否，1使用中
                creatorName:
                    type: string
                    description: 创建者
                updateTime:
                    type: integer
                    description: 更新时间
                    format: int64
        api.scenemanage.v1.SceneTemplateListReply:
            type: object
            properties:
                content:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.scenemanage.v1.SceneTemplate'
                    description: 场景模版列表
                totalElements:
                    type: integer
                    description: 字典模版总个数
                    format: int32
...